+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ## Fixed
 - Fix broken search in WebhookEventAdmin
+- Fix overriding "end_of_period" for canceled billing agreements
 
 ## [0.15.0] - 2020-01-30
 ## Added